pyproject.toml
[project]
name = "dhg503-project"
version = "0.1.0"
description = "DH503 Project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.13",
    "asyncio>=3.4.3",
    "google-cloud-texttospeech>=2.25.0",
    "jsonschema>=4.23.0",
    "lxml>=5.3.1",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "tqdm>=4.67.1",
    "youtube-dl>=2021.12.17",
    "ytdl>=1.0.0rc5",
]
